requirements.txt
pyyaml==6.0
matplotlib==3.5.3
absl-py==1.0.0
addict==2.4.0
affine==2.3.1
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
alembic==1.7.7
antlr4-python3-runtime==4.8
anyio==3.4.0
appdirs==1.4.4
argcomplete==1.12.3
asgiref==3.5.0
async-timeout==4.0.1
asynctest==0.13.0
Babel==2.9.1
bbox-utility==1.0.13
bitarray==2.3.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.6.15
click==8.1.2
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
colorama==0.4.4
coloredlogs==15.0.1
commonmark==0.9.1
cucim==22.4.0
cupy-cuda113==10.5.0
Cython==0.29.26
databricks-cli==0.16.6
datasets @ git+https://github.com/huggingface/datasets.git@855ae00330f9896bc308f01b302b2266a97e5e77
dill==0.3.4
Django==3.2.12
docker==5.0.3
docker-pycreds==0.4.0
efficientnet-pytorch==0.6.3
einops==0.2.0
et-xmlfile==1.1.0
faiss-gpu==1.7.2
fastjsonschema==2.15.3
fastrlock==0.8
ffmpeg==1.4
filelock==3.6.0
Flask==2.1.1
flatbuffers==2.0
fonttools==4.28.2
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
gdown==4.4.0
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.6.0
huggingface-hub==0.5.1
humanfriendly==10.0
hydra-core==1.0.7
image==1.5.33
imagecodecs==2021.11.20
imageio==2.14.1
imgaug==0.4.0
importlib-resources==5.4.0
install==1.3.5
itk==5.2.1.post1
itk-core==5.2.1.post1
itk-filtering==5.2.1.post1
itk-io==5.2.1.post1
itk-numerics==5.2.1.post1
itk-registration==5.2.1.post1
itk-segmentation==5.2.1.post1
itsdangerous==2.1.2
jiwer==2.3.0
json5==0.9.6
llvmlite==0.38.0
lmdb==1.0.0
loguru==0.6.0
Mako==1.2.0
Markdown==3.3.6
mkl-fft==1.3.1
mkl-service==2.4.0
mlflow==1.25.1
mmcv==1.5.0
monai==0.8.1
mpmath==1.2.1
multidict==5.2.0
multiprocess==0.70.12.2
munch==2.5.0
munkres==1.1.4
natsort==8.1.0
nbclassic==0.3.4
nbclient==0.6.3
networkx==2.6.3
nibabel==3.2.2
nltk==3.7
npm==0.1.1
nvidia-ml-py==11.450.51
oauthlib==3.1.1
omegaconf==2.0.6
onnxruntime==1.12.1
openpyxl==3.0.9
openslide-python==1.1.2
optional-django==0.1.0
pathtools==0.1.2
Pillow==9.2.0
plotly==5.4.0
ply==3.11
portalocker==2.3.2
prefetch-generator==1.0.1
pretrainedmodels==0.7.4
prometheus-flask-exporter==0.20.1
promise==2.3
prompt-toolkit==3.0.23
protobuf==3.19.1
psutil==5.9.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyDeprecate==0.3.1
PyJWT==2.3.0
PyMatting==1.1.8
pystyle==2.0
python-bidi==0.4.2
python-Levenshtein==0.12.2
pytorch-ignite==0.4.8
pytorch-lightning==1.5.3
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
querystring-parser==1.2.4
rasterio==1.2.10
regex==2021.11.10
requests-oauthlib==1.3.0
resampy==0.2.2
responses==0.18.0
rsa==4.8
sacrebleu==2.0.0
sacremoses==0.0.49
scikit-image==0.19.1
scikit-learn==1.0
seaborn==0.11.2
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools-scm==6.3.2
Shapely==1.8.0
shortuuid==1.0.9
smart-open==6.0.0
smmap==5.0.0
sniffio==1.2.0
snuggs==1.4.7
sox==1.4.1
sympy==1.10.1
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
termcolor==1.1.0
thop==0.1.1.post2207130030
tifffile==2021.11.2
tokenizers==0.12.1
tomli==1.2.2
tqdm==4.64.0
transformers==4.18.0
ttach==0.0.3
Unidecode==1.3.4
wandb==0.12.18
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
xlrd==2.0.1
xxhash==3.0.0
yapf==0.32.0
yarl==1.7.2
vietocr==0.3.8
pandas
numpy